Move datamodule into depthcharge components
